moves card routes from server.js to their own api file, adds statuses bookshelf model for future use
