+++ server/database/decorator.js
@@ -1,10 +1,12 @@
 const User = require("./models/User");
 const Card = require("./models/Card");
+const Status = require("./models/Status");
 
 module.exports = function(req, res, next) {
   req.database = {
-    User: User,
-    Card: Card
+    User,
+    Card,
+    Status
   };
 
   next();
